requirements.txt
appdirs==1.4.4
asgiref==3.6.0
astroid==2.4.2
autopep8==1.5.4
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
# Brlapi==0.7.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
colorama==0.4.4
# command-not-found==0.3
crispy-bootstrap4==2022.1
cryptography==40.0.2
# cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.4
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
# distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.2.1
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-allauth==0.54.0
django-booking==0.7.2
django-countries==7.5.1
django-crispy-forms==2.0
django-datatables-too==0.1.4
django-environ==0.4.5
django-extra-views==0.14.0
django-filter==2.4.0
django-hvad==1.8.0
django-libs==2.0.3
django-shortcuts==1.6
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.10
importlib-metadata==1.5.0
kazam==1.5.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.5.0
pipreqs==0.4.13
protobuf==3.6.1
psycopg==3.1.8
psycopg2-binary==2.9.6
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.21
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.23
python3-openid==3.2.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
querystring-parser==1.2.4
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sqlparse==0.3.1
systemd-python==234
toml==0.10.2
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
whitenoise==5.2.0
wrapt==1.12.1
xkit==0.0.0
yarg==0.1.9
zipp==3.15.0
